requirements.txt
faker>=18.0.0
pyspark==3.3.3
Faker==18.13.0
apache-airflow
apache-airflow-providers-apache-spark